pyproject.toml
[tool.poetry]
name = "Astronomer"
version = "0.1.0"
description = ""
authors = ["Tomodachi94 <tomodachi94@pm.me>"]

[tool.poetry.dependencies]
python = "3.8"
python-dotenv = "^0.15.0"
"discord.py" = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
